move path functions to path.rs
